introduce a new run-time variable PNETCDF_VERBOSE_DEBUG_MODE that prints the location of error codes originzted
git-svn-id: https://svn.mcs.anl.gov/repos/parallel-netcdf/trunk@2196 c368ceb8-fd28-0410-a457-e46ee1d7cd4a

+++ test/testcases/buftype_freef.f
@@ -103,10 +103,10 @@
           ! define an MPI datatype using MPI_Type_create_subarray()
           one        = 1_MPI_OFFSET_KIND
           ghost      = 2
-          gsize(1)   = NX + 2 * ghost
-          gsize(2)   = NY + 2 * ghost
-          subsize(1) = NX
-          subsize(2) = NY
+          gsize(1)   = INT(NX) + 2 * ghost
+          gsize(2)   = INT(NY) + 2 * ghost
+          subsize(1) = INT(NX)
+          subsize(2) = INT(NY)
           a_start(1) = ghost - 1
           a_start(2) = ghost - 1
